Add a data model class
Add a dm (data model) namespace and use it for storing entered data.
Use hidden field with listener to data store that stores the current
search options as JSON. Use this fields data to restore the values of
multiselect field.

This architecture is modeled after what's in mw.rcfilters code