Avoid endless loops in MenuedInput
populateFromStore clears and sets tags in the input, which caused an
infinite loops because the changes caused change events (which triggred
store changes, etc).

Now populateFromStore does what it should - only change the value when
store and input vlaue differ (which is the case when the store value is
changed externally, e.g. by removing the pill from the preview).